don't transition to error state for fetch/etc. errors
The ERROR state is meant to indicate terminal errors (like the session
being over). authkit-js will not attempt to retrieve new tokens after
when in an ERROR state.

Temporary errors (like `fetch` or tab-lock errors) should _not_ move to
the error state. This allows future calls to `getAccessToken` to attempt
to retrieve a new token.